+++ test/exponential_family_tests.jl
@@ -1,83 +1,7 @@
-module KnownExponentialFamilyDistributionTest
 
-using ExponentialFamily, Distributions, Test, StatsFuns, BenchmarkTools, Random, FillArrays
+@testitem "pack_parameters" begin
+    include("./exponential_family_setuptests.jl")
 
-import Distributions: RealInterval, ContinuousUnivariateDistribution, Univariate
-import ExponentialFamily: basemeasure, sufficientstatistics, logpartition, insupport, ConstantBaseMeasure
-import ExponentialFamily: getnaturalparameters, getbasemeasure, getsufficientstatistics, getlogpartition, getsupport
-import ExponentialFamily: ExponentialFamilyDistributionAttributes, NaturalParametersSpace
-import ExponentialFamily: paramfloattype, convert_paramfloattype
-
-# import ExponentialFamily:
-#     ExponentialFamilyDistribution, getnaturalparameters, getconditioner, reconstructargument!, as_vec,
-#     pack_naturalparameters, unpack_naturalparameters, insupport
-# import Distributions: pdf, logpdf, cdf
-
-## ===========================================================================
-## Tests fixtures
-const ArbitraryExponentialFamilyAttributes = ExponentialFamilyDistributionAttributes(
-    (x) -> 1 / x,
-    ((x) -> x, (x) -> log(x)),
-    (η) -> 1 / sum(η),
-    RealInterval(0, Inf)
-)
-
-# Arbitrary distribution (un-conditioned)
-struct ArbitraryDistributionFromExponentialFamily <: ContinuousUnivariateDistribution
-    p1::Float64
-    p2::Float64
-end
-
-ExponentialFamily.isproper(::NaturalParametersSpace, ::Type{ArbitraryDistributionFromExponentialFamily}, η, conditioner) = isnothing(conditioner)
-ExponentialFamily.isbasemeasureconstant(::Type{ArbitraryDistributionFromExponentialFamily}) = ConstantBaseMeasure()
-ExponentialFamily.getbasemeasure(::Type{ArbitraryDistributionFromExponentialFamily}) = (x) -> oneunit(x)
-ExponentialFamily.getsufficientstatistics(::Type{ArbitraryDistributionFromExponentialFamily}) =
-    ((x) -> x, (x) -> log(x))
-ExponentialFamily.getlogpartition(::NaturalParametersSpace, ::Type{ArbitraryDistributionFromExponentialFamily}) = (η) -> 1 / sum(η)
-ExponentialFamily.getsupport(::Type{ArbitraryDistributionFromExponentialFamily}) = RealInterval(0, Inf)
-
-ExponentialFamily.vague(::Type{ArbitraryDistributionFromExponentialFamily}) =
-    ArbitraryDistributionFromExponentialFamily(1.0, 1.0)
-
-Distributions.params(dist::ArbitraryDistributionFromExponentialFamily) = (dist.p1, dist.p2)
-
-(::MeanToNatural{ArbitraryDistributionFromExponentialFamily})(params::Tuple) = (params[1] + 1, params[2] + 1)
-(::NaturalToMean{ArbitraryDistributionFromExponentialFamily})(params::Tuple) = (params[1] - 1, params[2] - 1)
-
-ExponentialFamily.unpack_parameters(::Type{ArbitraryDistributionFromExponentialFamily}, η) = (η[1], η[2])
-
-# Arbitrary distribution (conditioned)
-struct ArbitraryConditionedDistributionFromExponentialFamily <: ContinuousUnivariateDistribution
-    con::Int
-    p1::Float64
-end
-
-ExponentialFamily.isproper(::NaturalParametersSpace, ::Type{ArbitraryConditionedDistributionFromExponentialFamily}, η, conditioner) = isinteger(conditioner)
-ExponentialFamily.isbasemeasureconstant(::Type{ArbitraryConditionedDistributionFromExponentialFamily}) = NonConstantBaseMeasure()
-ExponentialFamily.getbasemeasure(::Type{ArbitraryConditionedDistributionFromExponentialFamily}, conditioner) = (x) -> x^conditioner
-ExponentialFamily.getsufficientstatistics(::Type{ArbitraryConditionedDistributionFromExponentialFamily}, conditioner) =
-    ((x) -> log(x - conditioner),)
-ExponentialFamily.getlogpartition(::NaturalParametersSpace, ::Type{ArbitraryConditionedDistributionFromExponentialFamily}, conditioner) =
-    (η) -> conditioner / sum(η)
-ExponentialFamily.getsupport(::Type{ArbitraryConditionedDistributionFromExponentialFamily}) = RealInterval(0, Inf)
-
-ExponentialFamily.vague(::Type{ArbitraryConditionedDistributionFromExponentialFamily}) =
-    ArbitraryConditionedDistributionFromExponentialFamily(1.0, -2)
-
-Distributions.params(dist::ArbitraryConditionedDistributionFromExponentialFamily) = (dist.con, dist.p1)
-
-ExponentialFamily.separate_conditioner(::Type{ArbitraryConditionedDistributionFromExponentialFamily}, params) = ((params[2],), params[1])
-ExponentialFamily.join_conditioner(::Type{ArbitraryConditionedDistributionFromExponentialFamily}, cparams, conditioner) = (conditioner, cparams...)
-
-(::MeanToNatural{ArbitraryConditionedDistributionFromExponentialFamily})(params::Tuple, conditioner::Number) = (params[1] + conditioner,)
-(::NaturalToMean{ArbitraryConditionedDistributionFromExponentialFamily})(params::Tuple, conditioner::Number) = (params[1] - conditioner,)
-
-ExponentialFamily.unpack_parameters(::Type{ArbitraryConditionedDistributionFromExponentialFamily}, η) = (η[1],)
-
-## ===========================================================================
-## Tests
-
-@testset "pack_parameters" begin
     import ExponentialFamily: pack_parameters
 
     to_test_fixed = [
@@ -122,7 +46,9 @@ ExponentialFamily.unpack_parameters(::Type{ArbitraryConditionedDistributionFromE
     end
 end
 
-@testset "ExponentialFamilyDistributionAttributes" begin
+@testitem "ExponentialFamilyDistributionAttributes" begin
+    include("./exponential_family_setuptests.jl")
+
     @testset "getmapping" begin
         @test @inferred(getmapping(MeanParametersSpace() => NaturalParametersSpace(), ArbitraryDistributionFromExponentialFamily)) ===
               MeanToNatural{ArbitraryDistributionFromExponentialFamily}()
@@ -132,7 +58,6 @@ end
         @test @allocated(getmapping(NaturalParametersSpace() => MeanParametersSpace(), ArbitraryDistributionFromExponentialFamily)) === 0
     end
 
-    # See the `ArbitraryExponentialFamilyAttributes` defined in the fixtures (above)
     @testset let attributes = ArbitraryExponentialFamilyAttributes
         @test @inferred(getbasemeasure(attributes)(2.0)) ≈ 0.5
         @test @inferred(getsufficientstatistics(attributes)[1](2.0)) ≈ 2.0
@@ -183,9 +108,9 @@ end
     end
 end
 
-@testset "ExponentialFamilyDistribution" begin
+@testitem "ArbitraryDistributionFromExponentialFamily" begin
+    include("./exponential_family_setuptests.jl")
 
-    # See the `ArbitraryDistributionFromExponentialFamily` defined in the fixtures (above)
     @testset for member in (
         ExponentialFamilyDistribution(ArbitraryDistributionFromExponentialFamily, [2.0, 2.0]),
         convert(ExponentialFamilyDistribution, ArbitraryDistributionFromExponentialFamily(1.0, 1.0))
@@ -241,6 +166,8 @@ end
         # This is important, because the generic prod version should simply call the in-place version
         @test @allocated(prod(ClosedProd(), member, member)) <= @allocated(similar(member))
         @test @allocated(prod(GenericProd(), member, member)) <= @allocated(similar(member))
+
+        # This test is actually passing, but does not work if you re-run tests for some reason (which is hapenning often during development)
         # @test @allocated(prod(PreserveTypeProd(ExponentialFamilyDistribution), member, member)) <=
         #       @allocated(similar(member))
 
@@ -255,9 +182,10 @@ end
         # Test that the generic in-place prod! version does not allocate at all
         @test @allocated(prod!(_similar, member, member)) === 0
     end
+end
 
-    @test @inferred(vague(ExponentialFamilyDistribution{ArbitraryDistributionFromExponentialFamily})) isa
-          ExponentialFamilyDistribution{ArbitraryDistributionFromExponentialFamily}
+@testitem "ArbitraryConditionedDistributionFromExponentialFamily" begin
+    include("./exponential_family_setuptests.jl")
 
     # See the `ArbitraryDistributionFromExponentialFamily` defined in the fixtures (above)
     # p1 = 3.0, con = -2
@@ -314,8 +242,14 @@ end
         end
     end
 
-    @test @inferred(vague(ExponentialFamilyDistribution{ArbitraryConditionedDistributionFromExponentialFamily})) isa
-          ExponentialFamilyDistribution{ArbitraryConditionedDistributionFromExponentialFamily}
 end
 
+@testitem "vague" begin
+    include("./exponential_family_setuptests.jl")
+
+    @test @inferred(vague(ExponentialFamilyDistribution{ArbitraryDistributionFromExponentialFamily})) isa
+          ExponentialFamilyDistribution{ArbitraryDistributionFromExponentialFamily}
+
+    @test @inferred(vague(ExponentialFamilyDistribution{ArbitraryConditionedDistributionFromExponentialFamily})) isa
+          ExponentialFamilyDistribution{ArbitraryConditionedDistributionFromExponentialFamily}
 end
